refactor(taskflow): extract feature.board
