+++ examples/taskflow/composeApp/src/commonMain/kotlin/org/reduxkotlin/sample/taskflow/action/Actions.kt
@@ -1,34 +1,8 @@
 package org.reduxkotlin.sample.taskflow.action
 
-import kotlinx.collections.immutable.PersistentSet
-import org.reduxkotlin.sample.taskflow.core.AccountId
 import org.reduxkotlin.sample.taskflow.core.Action
-import org.reduxkotlin.sample.taskflow.core.Board
-import org.reduxkotlin.sample.taskflow.core.BoardId
-import org.reduxkotlin.sample.taskflow.core.Card
 import org.reduxkotlin.sample.taskflow.core.CardId
-import org.reduxkotlin.sample.taskflow.core.ColumnId
-import org.reduxkotlin.sample.taskflow.core.LabelId
 import org.reduxkotlin.sample.taskflow.core.Route
-import kotlin.time.Instant
-
-// --- bot (server-truth; NOT undoable, no revert) ---
-data class BotMovedCard(val cardId: CardId, val to: ColumnId, val toIndex: Int) : Action
-
-data class BotAddedCard(val columnId: ColumnId, val card: Card) : Action
-
-// --- board lifecycle ---
-data object BoardClosed : Action
-
-data class BoardRestored(val board: Board) : Action
-
-data class StartCreateCard(val columnId: ColumnId) : Action
-
-data object CancelCreateCard : Action
-
-data class AddColumn(val id: ColumnId, val title: String) : Action
-
-// CreateBoard moved to …feature.boardlist.BoardListActions
 
 // --- nav ---
 
@@ -61,33 +35,6 @@ data class OpenCard(val cardId: CardId) : Action
 data object CloseCard : Action
 
 // Undo, Redo, PushUndo, SetUndoModel moved to …feature.undo.UndoActions
-
-// --- filter ---
-data class SetFilterQuery(val query: String) : Action
-
-data class SetFilterAssignee(val accountId: AccountId?) : Action
-
-data class ToggleFilterLabel(val labelId: LabelId) : Action
-
-// --- board/account loads (per-account) ---
-data class LoadBoardRequested(val boardId: BoardId) : Action
-
-data class LoadBoardSucceeded(val board: Board) : Action
-
-data class LoadBoardFailed(val boardId: BoardId, val error: String) : Action
-
-// LoadBoardListRequested, LoadBoardListSucceeded, LoadBoardListFailed moved to …feature.boardlist.BoardListActions
-
-// --- sync (per-account) — folded into SyncModel by syncReducer ---
-data class SyncStatusChanged(
-    val online: Boolean,
-    val pendingCount: Int,
-    val inFlight: PersistentSet<CardId>,
-    val lastSyncedAt: Instant?,
-    val lastError: String?,
-) : Action
-
-data object Refresh : Action
 
 // --- profile / activity (per-account) ---
 // EditProfile moved to …feature.account.AccountActions
