+++ examples/taskflow/composeApp/src/commonMain/kotlin/org/reduxkotlin/sample/taskflow/feature/board/AttachmentChip.kt
@@ -1,4 +1,4 @@
-package org.reduxkotlin.sample.taskflow.ui.components
+package org.reduxkotlin.sample.taskflow.feature.board
 
 import androidx.compose.animation.core.RepeatMode
 import androidx.compose.animation.core.animateFloat
