+++ examples/taskflow/composeApp/src/commonMain/kotlin/org/reduxkotlin/sample/taskflow/feature/board/BoardSliceModels.kt
@@ -1,4 +1,4 @@
-package org.reduxkotlin.sample.taskflow.model
+package org.reduxkotlin.sample.taskflow.feature.board
 
 import kotlinx.collections.immutable.PersistentSet
 import kotlinx.collections.immutable.persistentSetOf
